Changed valid registration policy from model to policy and moved to appropriate folder
